feat: derive stats from leaderboard
